Get rid of expensifyAPI
