+++ src/libs/actions/PersonalDetails.js
@@ -5,7 +5,7 @@ import ONYXKEYS from '../../ONYXKEYS';
 import md5 from '../md5';
 import CONST from '../../CONST';
 import NetworkConnection from '../NetworkConnection';
-import expensifyAPI from '../expensifyAPI';
+import API from '../API';
 
 let currentUserEmail;
 Onyx.connect({
@@ -87,7 +87,7 @@ function formatPersonalDetails(personalDetailsList) {
  * Get the timezone of the logged in user
  */
 function fetchTimezone() {
-    expensifyAPI.Get({
+    API.Get({
         returnValueList: 'nameValuePairs',
         name: 'timeZone',
     })
@@ -104,7 +104,7 @@ function fetchTimezone() {
  * Get the personal details for our organization
  */
 function fetch() {
-    expensifyAPI.Get({
+    API.Get({
         returnValueList: 'personalDetailsList',
     })
         .then((data) => {
@@ -134,7 +134,7 @@ function fetch() {
  * @param {String} emailList
  */
 function getForEmails(emailList) {
-    expensifyAPI.PersonalDetails_GetForEmails({emailList})
+    API.PersonalDetails_GetForEmails({emailList})
         .then((data) => {
             const details = _.pick(data, emailList.split(','));
             Onyx.merge(ONYXKEYS.PERSONAL_DETAILS, formatPersonalDetails(details));
