+++ src/pages/home/report/ReportActionItemFragment.js
@@ -10,7 +10,7 @@ import styles, {webViewStyles, colors} from '../../../styles/StyleSheet';
 import Text from '../../../components/Text';
 import AnchorForCommentsOnly from '../../../components/AnchorForCommentsOnly';
 import InlineCodeBlock from '../../../components/InlineCodeBlock';
-import {getAuthToken} from '../../../libs/expensifyAPI';
+import API from '../../../libs/API';
 import ImageThumbnailWithModal from '../../../components/ImageThumbnailWithModal';
 import Config from '../../../CONFIG';
 
@@ -70,11 +70,11 @@ class ReportActionItemFragment extends React.PureComponent {
             img: (htmlAttribs, children, convertedCSSStyles, passProps) => {
                 // Attaches authTokens as a URL parameter to load image attachments
                 let previewSource = htmlAttribs['data-expensify-source']
-                    ? `${htmlAttribs.src}?authToken=${getAuthToken()}`
+                    ? `${htmlAttribs.src}?authToken=${API.getAuthToken()}`
                     : htmlAttribs.src;
 
                 let source = htmlAttribs['data-expensify-source']
-                    ? `${htmlAttribs['data-expensify-source']}?authToken=${getAuthToken()}`
+                    ? `${htmlAttribs['data-expensify-source']}?authToken=${API.getAuthToken()}`
                     : htmlAttribs.src;
 
                 // Update the image URL so the images can be accessed depending on the config environment
